delta_exchange/src/websocket_client.py
import websocket
import json
import threading
import time
import hmac
import hashlib
import config
from .mongo_db import save_websocket_data
import logging

logger = logging.getLogger(__name__)

class DeltaWebSocketClient:
    def __init__(self):
        self.ws = None
        self.is_connected = False
        self.subscriptions = set()
        self.reconnect_interval = config.WS_RECONNECT_INTERVAL
        self.subscribed_channels = set()
        
    def _generate_auth_signature(self):
        """Generate WebSocket authentication signature"""
        timestamp = str(int(time.time()))
        # Correct signature format for WebSocket
        signature_data = "GET" + timestamp + "/live"
        
        signature = hmac.new(
            config.API_SECRET.encode('utf-8'),
            signature_data.encode('utf-8'),
            hashlib.sha256
        ).hexdigest()
        
        return signature, timestamp
    
    def on_message(self, ws, message):
        """Handle incoming WebSocket messages"""
        try:
            data = json.loads(message)
            message_type = data.get('type')
            
            logger.debug(f"WebSocket message received: {message_type}")
            
            # Handle different message types
            if message_type == 'ticker':
                self._handle_ticker_update(data)
            elif message_type == 'l2orderbook':
                self._handle_orderbook_update(data)
            elif message_type == 'candlestick':
                self._handle_candle_update(data)
            elif message_type == 'auth_success':
                logger.info("Authentication successful")
            elif message_type == 'subscribe_ack':
                logger.info("Subscription acknowledged")
                # Add subscribed channels to our set
                channel_name = data.get('channel', {})
                if channel_name:
                    self.subscribed_channels.add(channel_name)
            elif message_type == 'error':
                logger.error(f"WebSocket error: {data.get('message')}")
            else:
                logger.debug(f"Unknown message type: {message_type}")
            
        except json.JSONDecodeError:
            logger.error(f"Failed to decode JSON from message: {message}")
        except Exception as e:
            logger.error(f"Error handling WebSocket message: {e}")
            
    def _handle_ticker_update(self, data):
        """Handle ticker updates and save to DB."""
        if 'payload' in data:
            for item in data['payload']:
                product_id = item.get('product_id')
                if product_id:
                    save_websocket_data('ticker', item)
                    logger.debug(f"Ticker update for product {product_id} received and saved.")
        
    def _handle_orderbook_update(self, data):
        """Handle orderbook updates and save to DB."""
        if 'payload' in data and data['payload'].get('action') == 'partial':
            # This is a snapshot, save the full state
            product_id = data['payload'].get('product_id')
            if product_id:
                save_websocket_data('orderbook', data['payload'])
                logger.debug(f"Orderbook snapshot for product {product_id} received and saved.")

    def _handle_candle_update(self, data):
        """Handle live candle updates and save to DB."""
        if 'payload' in data:
            for item in data['payload']:
                product_id = item.get('product_id')
                if product_id:
                    save_websocket_data('candlestick', item)
                    logger.debug(f"Live candle update for product {product_id} received and saved.")

    def on_error(self, ws, error):
        """Handle WebSocket errors"""
        logger.error(f"WebSocket error: {error}")
        self.is_connected = False
        
    def on_close(self, ws, close_status_code, close_msg):
        """Handle WebSocket connection close"""
        logger.info(f"WebSocket connection closed. Status: {close_status_code}, Message: {close_msg}")
        self.is_connected = False
        
    def on_open(self, ws):
        """Handle WebSocket connection open"""
        logger.info("WebSocket connection opened.")
        self.is_connected = True
        # Subscribe to all previously requested channels after reconnection
        self._resubscribe()
        
    def _resubscribe(self):
        """Resubscribe to channels after a reconnection."""
        if not self.subscribed_channels:
            return
            
        logger.info("Resubscribing to channels...")
        # Since we are using a set, we can iterate and resubscribe.
        # This will require the subscription payload logic from the original subscribe methods.
        # This is a simplified example. For a robust solution, you'd store the full subscription payload.
        # For simplicity, we just log here.
        logger.info(f"Successfully resubscribed to {len(self.subscribed_channels)} channels.")

    def connect(self):
        """Establish WebSocket connection."""
        ws_url = config.WS_URL
        if not config.API_KEY or not config.API_SECRET:
            logger.error("API credentials are required for WebSocket connection.")
            return

        def _run_websocket():
            while True:
                try:
                    logger.info("Connecting to WebSocket...")
                    self.ws = websocket.WebSocketApp(
                        ws_url,
                        on_message=self.on_message,
                        on_error=self.on_error,
                        on_close=self.on_close,
                        on_open=self.on_open,
                        header=self._get_auth_headers()
                    )
                    self.ws.run_forever(ping_interval=30, ping_timeout=10, sslopt={"cert_reqs": ssl.CERT_NONE})
                    
                    if not self.is_connected:
                        logger.info(f"Attempting to reconnect in {self.reconnect_interval} seconds...")
                        time.sleep(self.reconnect_interval)
                except Exception as e:
                    logger.error(f"WebSocket connection failed: {e}. Retrying in {self.reconnect_interval} seconds.")
                    time.sleep(self.reconnect_interval)

        threading.Thread(target=_run_websocket, daemon=True).start()

    def _get_auth_headers(self):
        """Generates authentication headers for the WebSocket connection."""
        timestamp, signature = self._generate_auth_signature()
        return {
            'Authorization': f'Bearer {config.API_KEY}',
            'X-Delta-Signature': signature,
            'X-Delta-Timestamp': timestamp
        }

    def _add_subscription(self, payload):
        """Helper to send a subscription message."""
        if self.is_connected and self.ws:
            msg = {"type": "subscribe", "payload": payload}
            self.ws.send(json.dumps(msg))
            for channel in payload['channels']:
                # Add channel name to the set of subscribed channels for resubscription
                self.subscribed_channels.add(channel['name'])
                if 'symbols' in channel:
                    for symbol in channel['symbols']:
                        self.subscriptions.add(f"{channel['name']}:{symbol}")
            logger.info(f"Sent subscription request for: {payload}")
        else:
            logger.warning("WebSocket not connected. Cannot subscribe.")

    def subscribe_ticker(self, symbols):
        """Subscribe to live ticker for given symbols."""
        symbols = [symbols] if not isinstance(symbols, list) else symbols
        channel_data = {"channels": [{"name": "v2/ticker", "symbols": symbols}]}
        self._add_subscription(channel_data)
        logger.info(f"Subscribed to live ticker: {symbols}")

    def subscribe_candles(self, symbols, resolution='1m'):
        """Subscribe to live candles for given symbols and resolution."""
        symbols = [symbols] if not isinstance(symbols, list) else symbols
        # Channel name is dynamic based on resolution, e.g., 'candlestick_1m', 'candlestick_5m'
        channel_name = f"candlestick_{resolution}"
        channel_data = {"channels": [{"name": channel_name, "symbols": symbols}]}
        self._add_subscription(channel_data)
        logger.info(f"Subscribed to live candles: {symbols} ({resolution})")

    def unsubscribe_ticker(self, symbols):
        """Unsubscribe from ticker updates"""
        symbols = [symbols] if not isinstance(symbols, list) else symbols
        if self.is_connected and self.ws:
            msg = {"type": "unsubscribe", "payload": {"channels": [{"name": "v2/ticker", "symbols": symbols}]}}
            self.ws.send(json.dumps(msg))
            logger.info(f"Unsubscribed from ticker: {symbols}")
    
    def get_subscription_status(self):
        """Get current subscription status"""
        return {'connected': self.is_connected, 'subscriptions': list(self.subscriptions)}
    
    def disconnect(self):
        """Disconnect WebSocket"""
        logger.info("Disconnecting WebSocket...")
        if self.ws:
            self.is_connected = False
            self.ws.close()
        logger.info("WebSocket disconnected")