delta_exchange/requirements.txt
# Core requirements for Delta Exchange Trading System
requests==2.31.0
pymongo==4.6.0
websocket-client==1.6.4
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Optional but recommended
python-dotenv==1.0.0
pydantic==2.5.0